+++ src/misc.h
@@ -12,18 +12,18 @@
 typedef unsigned char BYTE;
 typedef uint64_t WORD;
 
-extern const short WORD_SIZE;
-extern const short WORD_SIZE_BITS;
-extern const short DWORD_SIZE;
+extern const WORD WORD_SIZE;
+extern const WORD WORD_SIZE_BITS;
+extern const WORD DWORD_SIZE;
 
 #define ROTL64(X,C) (((X)<<(C))|((X)>>(64-(C))))
 #define ROTR64(X,C) (((X)>>(C))|((X)<<(64-(C))))
 
 enum mode { ENCRYPTION, DECRYPTION, KEYGEN };
 
-BYTE *keygen(int);
+BYTE *keygen(WORD);
 WORD pow_word(WORD, WORD);
 WORD get_file_size(FILE*);
 WORD align(WORD);
 WORD pad(WORD);
-int flush_data(FILE*, WORD*, WORD, enum mode);
+WORD flush_data(FILE*, WORD*, WORD, enum mode);
